added more modules, removed implicit imports
