+++ pIChemiSt/.flake8
@@ -3,3 +3,5 @@ exclude =
     legacy
     cli.py
     dimorphite.py
+    acd.py
+    charges.py
