+++ pIChemiSt/pichemist/cli.py
@@ -2,21 +2,20 @@ import sys, os, csv
 import argparse
 import json
 import os.path
-import subprocess
 
 # TODO: Explicit imports 
-from numpy import *
-from matplotlib.pyplot import *
+import numpy as np
+import matplotlib.pyplot as plt
 from itertools import cycle
 from rdkit import Chem
 from rdkit import RDLogger
 
 from pichemist.config import PKA_SETS_NAMES
+from pichemist.api import calc_pkas
 from pichemist.io import generate_input
 from pichemist.molecule import MolStandardiser
 from pichemist.molecule import PeptideCutter
 from pichemist.fasta.matcher import get_aa_pkas_for_list
-from pichemist.pkamatcher import PKaMatcher
 from pichemist.model import PKaMethod
 from pichemist.model import InputFormat
 from pichemist.model import OutputFormat
@@ -32,117 +31,6 @@ log = get_logger(__name__)
 RDLogger.DisableLog("rdApp.info")
 
 
-def clean_up_output(text):
-    txt=text.split('\n')
-    text_new=''
-    for line in txt: 
-        if "mkdir: cannot create directory '/.local': Permission denied" in line: continue
-        text_new += line+'\n'
-    return text_new
-
-def get_status_output(*args, **kwargs):
-    #from subprocess import Popen
-    p = subprocess.Popen(*args, **kwargs)
-    stdout, stderr = p.communicate()
-    return p.returncode, stdout, stderr
-
-def run_exe(exe):
-    #if sys.version_info[0] < 3:
-    #    from commands import getstatusoutput
-    #    status, output = getstatusoutput(exe)
-    #else:
-    status, stdout, stderr = get_status_output(exe.split(), stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-    output =  stdout.decode("utf-8")
-    if status != 0: raise Exception("ERROR: happend while executing "+exe+" : " + output)
-    output = clean_up_output(output)
-    return output
-    
-
-def calc_pkas_acdlabs(smi_list):
-    # use ACDlabs.
-    # perceptabat executable should be in the path.
-
-    pka_lim_base_1=2
-    pka_lim_base_2=15
-    pka_lim_acid_1=-5
-    pka_lim_acid_2=12
-
-    tmpfile='TMP_SMI_FOR_PKA.smi'
-    #tmpfile4='TMP_SMI_FOR_PKA.json'
-    if os.path.isfile(tmpfile): os.remove(tmpfile)
-    with open(tmpfile,'w') as smifile:
-        i=0
-        for smi in smi_list:
-            i+=1
-            smifile.write(smi+" tmpname"+str(i)+'\n')
-
-    tmpfile2='TMP_CLAB_OUTPUT.txt'
-    if os.path.isfile(tmpfile2): os.remove(tmpfile2)
-    #exe = 'perceptabat -TFNAME'+tmpfile2+' -MPKAAPP -TPKA ' + tmpfile
-    exe = 'perceptabat -TFNAME'+tmpfile2+' -MPKAAPPGALAS -TPKA ' + tmpfile
-    tmpoutput = run_exe(exe)
-    if not os.path.isfile(tmpfile2): 
-        print("ERROR: no tmpfile generated by acdlabs "+tmpfile2)
-        sys.exit(1)
-
-
-    with open(tmpfile2,'r') as f:
-        base_pkas=[]
-        acid_pkas=[]
-        diacid_pkas=[]
-        D={}
-        f.readline() # skip first line
-        for line in f.readlines():
-            ln=line.split()
-            mol_idx = int(ln[0])
-            if mol_idx not in D.keys(): D[mol_idx]={}
-            #if 'ACD_pKa_Apparent' in ln[1]: D[mol_idx]['pka']=float(ln[2])
-            if 'ACD_pKa_Apparent' in ln[1]: pka = float(ln[2])
-            if 'ACD_pKa_DissType_Apparent' in ln[1]: 
-                if ln[2] in ['MB','B']:
-                    if pka > pka_lim_base_1 and pka < pka_lim_base_2: 
-                        base_pkas.append((pka,smi_list[mol_idx-1]))
-                if ln[2] in ['MA','A']:
-                    if pka > pka_lim_acid_1 and pka < pka_lim_acid_2: 
-                        acid_pkas.append((pka,smi_list[mol_idx-1]))
-
-### Examle of perceptabatch output 
-#PKA: Calculate apparent values using classic algorithm
-#1	 ID: 1
-#1	 ACD_pKa_IonicForm_Apparent: L
-#1	 ACD_pKa_Apparent_1: 10.689
-#1	 ACD_pKa_Error_Apparent_1: 0.10
-#1	 ACD_pKa_DissAtom_Apparent_1: 5
-#1	 ACD_pKa_DissType_Apparent_1: MB
-#1	 ACD_pKa_Equation_Apparent_1: HL/H+L
-#1	 ACD_pKa_All_Apparent_1: pKa(HL/H+L; 5) = 10.69+/-0.10
-#2	 ID: 2
-#2	 ACD_pKa_IonicForm_Apparent: L
-#...
-#4	 ID: 4
-#4	 ACD_pKa_Caution_Apparent: The structure does not contain ionization centers calculated by current version of ACD/pKa
-#5	 ID: 5
-#5	 ACD_pKa_IonicForm_Apparent: L
-#5	 ACD_pKa_Apparent_1: 10.679
-#5	 ACD_pKa_Error_Apparent_1: 0.10
-#5	 ACD_pKa_DissAtom_Apparent_1: 1
-#5	 ACD_pKa_DissType_Apparent_1: MB
-#5	 ACD_pKa_Equation_Apparent_1: HL/H+L
-#5	 ACD_pKa_All_Apparent_1: pKa(HL/H+L; 1) = 10.68+/-0.10
-#5	 ACD_pKa_Apparent_2: 9.334
-#5	 ACD_pKa_Error_Apparent_2: 0.10
-#5	 ACD_pKa_DissAtom_Apparent_2: 6
-#5	 ACD_pKa_DissType_Apparent_2: B 
-#5	 ACD_pKa_Equation_Apparent_2: H2L/H+HL
-#5	 ACD_pKa_All_Apparent_2: pKa(H2L/H+HL; 6) = 9.33+/-0.10
-
-    lClean=False
-    if lClean:     
-            if os.path.isfile(tmpfile): os.remove(tmpfile)
-            if os.path.isfile(tmpfile2): os.remove(tmpfile2)
-
-    return (base_pkas,acid_pkas,diacid_pkas)
-
 def calc_molecule_constant_charge(net_Qs):
     q = [ v[0] for v in net_Qs]
     if len(q)>0:
@@ -151,44 +39,6 @@ def calc_molecule_constant_charge(net_Qs):
         constant_q = 0.0
 
     return constant_q
-
-def calc_net_Qs(smi_list):
-    net_Qs = []
-    for smi in smi_list:
-        mol = Chem.MolFromSmiles(smi)
-        mol = MolStandardiser().standardise_molecule(mol)
-
-        # positively charged N, but not connected to any negatively charged atom. To avoid azido and nitro groups being counted.
-        pattern = Chem.MolFromSmarts("[#7+;!H1;!H2;!H3;!H4!$([#7+]~[*-])]")
-        at_matches = mol.GetSubstructMatches(pattern)
-
-        at_matches_list = [y[0] for y in at_matches]
-
-        for v in at_matches_list:
-            net_Qs.append( (1,smi) )
-    
-    return net_Qs
-        
-        
-
-# calcualtes pKa for the list of smiles. 
-def calc_pkas(smi_list, method):
-
-    if method not in MODELS[PKaMethod]:
-        raise Exception("TODO...")
-
-    if method == PKaMethod.ACD.value:
-        base_pkas,acid_pkas,diacid_pkas = calc_pkas_acdlabs(smi_list)
-
-    if method == PKaMethod.PKA_MATCHER.value:
-        base_pkas,acid_pkas,diacid_pkas = PKaMatcher().calculate_pka_from_list(smi_list)
-
-    #if use_dimorphite:
-    #    base_pkas,acid_pkas,diacid_pkas = calc_pkas_dimorphite_dl(smi_list)
-
-    net_Qs = calc_net_Qs(smi_list) 
-    
-    return (base_pkas,acid_pkas,diacid_pkas,net_Qs)
 
 
         
@@ -271,7 +121,7 @@ def calculateIsoelectricPoint(base_pkas, acid_pkas, diacid_pkas, constant_q=0):
 def CalcChargepHCurve(base_pkas, acid_pkas, diacid_pkas, constant_q=0):
     pH_lim = define_pH_span()
     dpH=0.1
-    pH_a=arange(pH_lim[0],pH_lim[1]+dpH,dpH)
+    pH_a = np.arange(pH_lim[0],pH_lim[1]+dpH,dpH)
     Q_a=pH_a*0.0    
     for i in range(len(pH_a)):
         Q = calculateMolCharge(base_pkas, acid_pkas, diacid_pkas, pH_a[i],constant_q=constant_q)
@@ -298,51 +148,51 @@ def print_output_prop_dict(prop_dict, prop, print_pka_set=False):
         print(k.rjust(lj)  + "  " +  str(round(p,2)).ljust(lj) )
     print(" ")
 
-    if print_pka_set: print_pka_set()
+    if print_pka_set: print_pka_set()   # ANDREY: print_pka_set() does not exist anywhere
 
     return
 
 
 ### PLOT titration curve
-def plot_titration_curve(pH_Q_dict,figFileName):
-    matplotlib.rcParams.update({'font.size': 16})
+def _plot_titration_curve(pH_Q_dict,figFileName):
+    plt.matplotlib.rcParams.update({'font.size': 16})
     lines = ["-","--","-.",":"]
     w1=4.0 ; w2=3.0 ; w3=2.0 ; w4=1.0
     linew = [w1,w1, w2,w2, w3,3, w4,w4]
     linecycler = cycle(lines)
     linewcycler = cycle(linew)
 
-    figure(figsize=(8,6))
+    plt.figure(figsize=(8,6))
     i=0
     for pKaset in PKA_SETS_NAMES:
         i+=1
         pH_Q = pH_Q_dict[pKaset] 
-        l=plot(pH_Q[:,0],pH_Q[:,1],next(linecycler),label=pKaset,linewidth=next(linewcycler)) 
+        l = plt.plot(pH_Q[:,0],pH_Q[:,1],next(linecycler),label=pKaset,linewidth=next(linewcycler)) 
         if pKaset == 'IPC2_peptide': 
-            setp(l,linewidth=8,linestyle='-',color='k')
+            plt.setp(l,linewidth=8,linestyle='-',color='k')
 
         # Store data for output
         if i==1: 
             pH = pH_Q[:,0]
             Q_M = pH_Q[:,1]
         else:
-            Q_M = column_stack([Q_M,pH_Q[:,1]])
+            Q_M = np.column_stack([Q_M,pH_Q[:,1]])
 
-    plot(pH,pH*0,'k-')
-    plot([7,7],[np.min(Q_M),np.max(Q_M)],'k-')
+    plt.plot(pH,pH*0,'k-')
+    plt.plot([7,7],[np.min(Q_M),np.max(Q_M)],'k-')
     #xlim([np.min(pH),np.max(pH)])
-    xlim([2,12])
-    ylim([np.min(Q_M),np.max(Q_M)])
+    plt.xlim([2,12])
+    plt.ylim([np.min(Q_M),np.max(Q_M)])
 
-    legend(loc="center right", bbox_to_anchor=[1.1, 0.5],ncol=1, shadow=True, fontsize=10).get_frame().set_alpha(1)
-    ylabel('peptide charge')
-    xlabel('pH')
+    plt.legend(loc="center right", bbox_to_anchor=[1.1, 0.5],ncol=1, shadow=True, fontsize=10).get_frame().set_alpha(1)
+    plt.ylabel('peptide charge')
+    plt.xlabel('pH')
 	
-    minorticks_on()
-    grid(True)
+    plt.minorticks_on()
+    plt.grid(True)
 
     #show()
-    savefig(figFileName)
+    plt.savefig(figFileName)
     return
 
 # TODO: use input and type to determine input
@@ -360,6 +210,9 @@ def calc_pichemist(input_dict, method,
         mol = input_dict[mol_idx]['mol_obj']    
         mol = MolStandardiser().standardise_molecule(mol)
         frags_smi_list = PeptideCutter().break_amide_bonds_and_cap(mol)
+        # ANDREY: calc_net_Qs wants standardise_molecule() for each mol in the list
+        # should that be done here instead or maybe not at all since we do for the
+        # full peptide?
 
         # TODO: Match and calc pKas should be an API that contains the following
             # TODO: Match AA pKas from SMILES list should also be an API
@@ -388,12 +241,18 @@ def calc_pichemist(input_dict, method,
             all_base_pkas=[]
             all_acid_pkas=[]
             all_diacid_pkas=[]
-            if len(base_pkas) != 0: all_base_pkas,all_base_pkas_smi = zip(*base_pkas) 
-            else: all_base_pkas,all_base_pkas_smi = [],[]
-            if len(acid_pkas) != 0: all_acid_pkas,all_acid_pkas_smi = zip(*acid_pkas) 
-            else: all_acid_pkas,all_acid_pkas_smi = [],[]
-            if len(diacid_pkas) != 0: all_diacid_pkas,all_diacid_pkas_smi = zip(*diacid_pkas) 
-            else: all_diacid_pkas,all_diacid_pkas_smi = [],[]
+            if len(base_pkas) != 0:
+                all_base_pkas, all_base_pkas_smi = zip(*base_pkas) 
+            else:
+                all_base_pkas, all_base_pkas_smi = [],[]
+            if len(acid_pkas) != 0:
+                all_acid_pkas, all_acid_pkas_smi = zip(*acid_pkas) 
+            else:
+                all_acid_pkas, all_acid_pkas_smi = [],[]
+            if len(diacid_pkas) != 0:
+                all_diacid_pkas, all_diacid_pkas_smi = zip(*diacid_pkas) 
+            else:
+                all_diacid_pkas, all_diacid_pkas_smi = [],[]
 
             # calculate isoelectric point
             molecule_constant_charge = calc_molecule_constant_charge(net_Qs)
@@ -511,7 +370,7 @@ def calc_pichemist(input_dict, method,
         # plot titration curve
         if plot_titration_curve:
             figFileName = "OUT_titration_curve_pIChemiSt.png"
-            plot_titration_curve(pH_Q_dict,figFileName)
+            _plot_titration_curve(pH_Q_dict,figFileName)
         else:
             figFileName = ""
